ふかつどう: Modify Heads Up dismissal and hiding to be more intuitive
Stock behavior is swipe up to hide, and swipe left/right to dismiss.

Slim's behavior is to swipe right to hide, and swipe left to dismiss. This makes no sense and is hardly intuitive, the only way you wouldn't be confused is if you knew about it in the first place or somehow manage to figure it out yourself after some initial confusion.

The behavior of swiping downwards is kept from Slim, as it does make sense.