Modify Heads Up dismissal and hiding to be more intuitive
Stock behavior is dismiss regardless swiping up, left or right.

CM/Slim behavior is swipe right to hide and swipe left to dismiss. This makes absolutely no sense and is hardly intuitive, you would only know this was the case if you were previously informed or figured it out after some initial confusion.

This commit modifies the behavior to:
 - left/right: dismiss
 - up: hide
 - down: hide then open notification drawer so it appears there instead of above the notification drawer header